+++ tests/test_ensembles/test_direct_ensemble.py
@@ -206,9 +206,9 @@ def test_predict_with_return_components_fails(example_tsds, direct_ensemble_pipe
         direct_ensemble_pipeline.predict(ts=example_tsds, return_components=True)
 
 
-def test_params_to_tune_not_implemented(direct_ensemble_pipeline):
-    with pytest.raises(NotImplementedError, match="DirectEnsemble doesn't support this method"):
-        _ = direct_ensemble_pipeline.params_to_tune()
+def test_params_to_tune(direct_ensemble_pipeline):
+    result = direct_ensemble_pipeline.params_to_tune()
+    assert result == {}
 
 
 @pytest.mark.parametrize("n_jobs", (1, 5))